db get_create method, changed the way of parsing, change marshal fields

+++ resources/OfferResource.py
@@ -40,38 +40,11 @@ class OfferByUserIdResource(Resource):
         self.full_path = request.full_path
         self.args = main_parser.parse_args()
 
-        # User Data field parser
-        # Used for parsing the user and user meta fields inside the data field
-        self.user_data_parser = reqparse.RequestParser()
-        self.user_data_parser.add_argument('offer', type=dict, required=True, location=('data'))
-
-        # Usermeta parser
-        # Used for parsing the fields inside the user meta field
-        self.offer_parser = reqparse.RequestParser()
-        self.offer_parser.add_argument('subject_id', type=str, required=True, location=('offer',))
-        self.offer_parser.add_argument('level_id', type=str, required=True, location=('offer',))
-
     @api_validation
     @marshal_with(offer_fields)
     def get(self, id):
         user = session.query(User).filter(User.id == id).first()
         return user.offers, 200
-
-    @api_validation
-    @marshal_with(offer_fields)
-    def post(self, id):
-        data_args = self.data_parser.parse_args(self.args)
-        offer_args = self.offer_parser.parse_args(data_args)
-
-        user = session.query(User).filter(User.id == id).first()
-        if not user:
-            abort(404, message="User with id={} doesn't exist".format(id))
-
-        user.offers.append(Offer(subject_id=offer_args['subject_id'],
-                                 level_id=offer_args['level_id']))
-        session.add(user)
-        session.commit()
-        return user, 201
 
 
 class OfferResource(Resource):
@@ -94,37 +67,46 @@ class OfferResource(Resource):
         self.full_path = request.full_path
         self.args = main_parser.parse_args()
 
-        # Offer Search parser
-        # Used for parsing the search query arguments
-        self.offer_args_parser = reqparse.RequestParser()
-        self.offer_args_parser.add_argument('loc', type=str, required=True, location=('args'))
-        self.offer_args_parser.add_argument('range', type=int, required=True, location=('args',))
-        self.offer_args_parser.add_argument('subject', type=int, required=True, location=('args'))
-        self.offer_args_parser.add_argument('level', type=int, required=True, location=('args'))
-        self.offer_args_parser.add_argument('page', type=int, required=True, location=('args'))
-        self.offer_args_parser.add_argument('sortby', type=str, required=True, location=('args'))
-
     @api_validation
     @marshal_with(offer_fields)
     def get(self):
-        args = self.offer_args_parser.parse_args()
-        offers = session.query(Offer).filter(Offer.subject_id == args['subject'],
-                                             Offer.level_id == args['level']).all()
+        offer_query = offersearch_parser.parse_args()
+        offers = session.query(Offer).filter(Offer.subject_id == offer_query['subject'],
+                                             Offer.level_id == offer_query['level']).all()
 
-        loc_lat, loc_lon = map(float, args['loc'].split(','))
+        loc_lat, loc_lon = map(float, offer_query['loc'].split(','))
 
         result_offers = []
 
         for offer in offers:
             offer_lat = float(offer.user.meta.latitude)
             offer_lon = float(offer.user.meta.longitude)
-            if latlon_distance(loc_lat, loc_lon, offer_lat, offer_lon) < args['range']:
+            if latlon_distance(loc_lat, loc_lon, offer_lat, offer_lon) < offer_query['range']:
                 result_offers.append(offer)
 
         if not result_offers:
             abort(404, message="No offers found")
 
         return result_offers, 200
+
+    @api_validation
+    @authentication(None)
+    @marshal_with(offer_fields)
+    def post(self):
+        offer_args = offer_parser.parse_args(data_parser("offer", self.args))
+        loggedin_data = loggedin_parser.parse_args(data_parser("loggedin", self.args))
+
+        user = session.query(User).filter(User.id == loggedin_data['user_id']).first()
+        if not user:
+            abort(404, message="User with id={} doesn't exist".format(id))
+
+        offer = get_or_create(session, Offer, user_id=loggedin_data['user_id'],
+                              level_id=offer_args['level_id'],
+                              subject_id=offer_args['subject_id'])
+
+        session.add(offer)
+        session.commit()
+        return offer, 201
 
 
 class OfferByIdResource(Resource):
@@ -144,12 +126,18 @@ class OfferByIdResource(Resource):
         self.args = main_parser.parse_args()
 
     @api_validation
+    @authentication(None)
     @marshal_with(offer_fields)
     def delete(self, id):
+        loggedin_data = loggedin_parser.parse_args(data_parser("loggedin". self.args))
+
         offer = session.query(Offer).filter(Offer.id == id).first()
+        if (offer.user.id != loggedin_data['user_id']):
+            abort(401, message="Not authorized to delete offer with id={}".format(id))
+
         if not offer:
             abort(404, message="Offer with id={} doesn't exist".format(id))
 
         session.delete(offer)
         session.commit()
-        return {}, 201
+        return {}, 200
